refactor(types): Make strapi types common
